# internal/ports/openapi_api.gen.go
// Package ports provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.1 DO NOT EDIT.
package ports

import (
	"fmt"
	"net/url"

	"github.com/gofiber/fiber/v2"
	"github.com/oapi-codegen/runtime"
)

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Authenticate a user
	// (POST /api/auth/login)
	Login(c *fiber.Ctx) error
	// Log out the current user
	// (POST /api/auth/logout)
	Logout(c *fiber.Ctx) error
	// Refresh access token
	// (POST /api/auth/refresh)
	Refresh(c *fiber.Ctx) error
	// Register a new user.
	// (POST /api/auth/register)
	Register(c *fiber.Ctx) error
	// Redirect to Google OAuth page
	// (GET /api/oauth/google)
	GoogleOAuth(c *fiber.Ctx) error
	// Get google redirect page
	// (GET /api/oauth/google/callback)
	GoogleOAuthCb(c *fiber.Ctx, params GoogleOAuthCbParams) error
	// Download a specific user's image
	// (GET /api/user/download/image/{imageId})
	DownloadImage(c *fiber.Ctx, imageId ImageId) error
	// Process a user's image (SSE only)
	// (POST /api/user/process/image)
	ProcessImage(c *fiber.Ctx) error
	// Upload a user image
	// (POST /api/user/upload/image)
	UploadImage(c *fiber.Ctx) error
	// Get information about a specific user
	// (GET /api/user/{username})
	GetUserInfo(c *fiber.Ctx, username string) error
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler ServerInterface
}

type MiddlewareFunc fiber.Handler

// Login operation middleware
func (siw *ServerInterfaceWrapper) Login(c *fiber.Ctx) error {

	return siw.Handler.Login(c)
}

// Logout operation middleware
func (siw *ServerInterfaceWrapper) Logout(c *fiber.Ctx) error {

	c.Context().SetUserValue(Refresh_tokenScopes, []string{})

	return siw.Handler.Logout(c)
}

// Refresh operation middleware
func (siw *ServerInterfaceWrapper) Refresh(c *fiber.Ctx) error {

	c.Context().SetUserValue(Refresh_tokenScopes, []string{})

	return siw.Handler.Refresh(c)
}

// Register operation middleware
func (siw *ServerInterfaceWrapper) Register(c *fiber.Ctx) error {

	return siw.Handler.Register(c)
}

// GoogleOAuth operation middleware
func (siw *ServerInterfaceWrapper) GoogleOAuth(c *fiber.Ctx) error {

	return siw.Handler.GoogleOAuth(c)
}

// GoogleOAuthCb operation middleware
func (siw *ServerInterfaceWrapper) GoogleOAuthCb(c *fiber.Ctx) error {

	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params GoogleOAuthCbParams

	var query url.Values
	query, err = url.ParseQuery(string(c.Request().URI().QueryString()))
	if err != nil {
		return fiber.NewError(fiber.StatusBadRequest, fmt.Errorf("Invalid format for query string: %w", err).Error())
	}

	// ------------- Required query parameter "code" -------------

	if paramValue := c.Query("code"); paramValue != "" {

	} else {
		err = fmt.Errorf("Query argument code is required, but not found")
		c.Status(fiber.StatusBadRequest).JSON(err)
		return err
	}

	err = runtime.BindQueryParameter("form", true, true, "code", query, &params.Code)
	if err != nil {
		return fiber.NewError(fiber.StatusBadRequest, fmt.Errorf("Invalid format for parameter code: %w", err).Error())
	}

	// ------------- Optional query parameter "state" -------------

	err = runtime.BindQueryParameter("form", true, false, "state", query, &params.State)
	if err != nil {
		return fiber.NewError(fiber.StatusBadRequest, fmt.Errorf("Invalid format for parameter state: %w", err).Error())
	}

	return siw.Handler.GoogleOAuthCb(c, params)
}

// DownloadImage operation middleware
func (siw *ServerInterfaceWrapper) DownloadImage(c *fiber.Ctx) error {

	var err error

	// ------------- Path parameter "imageId" -------------
	var imageId ImageId

	err = runtime.BindStyledParameterWithOptions("simple", "imageId", c.Params("imageId"), &imageId, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		return fiber.NewError(fiber.StatusBadRequest, fmt.Errorf("Invalid format for parameter imageId: %w", err).Error())
	}

	c.Context().SetUserValue(Access_tokenScopes, []string{})

	return siw.Handler.DownloadImage(c, imageId)
}

// ProcessImage operation middleware
func (siw *ServerInterfaceWrapper) ProcessImage(c *fiber.Ctx) error {

	c.Context().SetUserValue(Access_tokenScopes, []string{})

	return siw.Handler.ProcessImage(c)
}

// UploadImage operation middleware
func (siw *ServerInterfaceWrapper) UploadImage(c *fiber.Ctx) error {

	c.Context().SetUserValue(Access_tokenScopes, []string{})

	return siw.Handler.UploadImage(c)
}

// GetUserInfo operation middleware
func (siw *ServerInterfaceWrapper) GetUserInfo(c *fiber.Ctx) error {

	var err error

	// ------------- Path parameter "username" -------------
	var username string

	err = runtime.BindStyledParameterWithOptions("simple", "username", c.Params("username"), &username, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		return fiber.NewError(fiber.StatusBadRequest, fmt.Errorf("Invalid format for parameter username: %w", err).Error())
	}

	c.Context().SetUserValue(Access_tokenScopes, []string{})

	return siw.Handler.GetUserInfo(c, username)
}

// FiberServerOptions provides options for the Fiber server.
type FiberServerOptions struct {
	BaseURL     string
	Middlewares []MiddlewareFunc
}

// RegisterHandlers creates http.Handler with routing matching OpenAPI spec.
func RegisterHandlers(router fiber.Router, si ServerInterface) {
	RegisterHandlersWithOptions(router, si, FiberServerOptions{})
}

// RegisterHandlersWithOptions creates http.Handler with additional options
func RegisterHandlersWithOptions(router fiber.Router, si ServerInterface, options FiberServerOptions) {
	wrapper := ServerInterfaceWrapper{
		Handler: si,
	}

	for _, m := range options.Middlewares {
		router.Use(fiber.Handler(m))
	}

	router.Post(options.BaseURL+"/api/auth/login", wrapper.Login)

	router.Post(options.BaseURL+"/api/auth/logout", wrapper.Logout)

	router.Post(options.BaseURL+"/api/auth/refresh", wrapper.Refresh)

	router.Post(options.BaseURL+"/api/auth/register", wrapper.Register)

	router.Get(options.BaseURL+"/api/oauth/google", wrapper.GoogleOAuth)

	router.Get(options.BaseURL+"/api/oauth/google/callback", wrapper.GoogleOAuthCb)

	router.Get(options.BaseURL+"/api/user/download/image/:imageId", wrapper.DownloadImage)

	router.Post(options.BaseURL+"/api/user/process/image", wrapper.ProcessImage)

	router.Post(options.BaseURL+"/api/user/upload/image", wrapper.UploadImage)

	router.Get(options.BaseURL+"/api/user/:username", wrapper.GetUserInfo)

}