# internal/ports/openapi_types.gen.go
// Package ports provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.1 DO NOT EDIT.
package ports

import (
	openapi_types "github.com/oapi-codegen/runtime/types"
)

const (
	Access_tokenScopes  = "access_token.Scopes"
	Refresh_tokenScopes = "refresh_token.Scopes"
)

// BasicUserInfo Basic information about a user
type BasicUserInfo struct {
	// Email users email
	Email *openapi_types.Email `json:"email,omitempty"`

	// Fullname users full name
	Fullname *string `json:"fullname,omitempty"`
	Id       *string `json:"id,omitempty"`

	// Images (array of users images)
	Images *[]ImageId `json:"images,omitempty"`

	// Username usersname
	Username *string `json:"username,omitempty"`
}

// Error defines model for Error.
type Error struct {
	Title *string `json:"title,omitempty"`
	Type  *string `json:"type,omitempty"`
}

// ImageId defines model for ImageId.
type ImageId = string

// ImageProcessingOptions defines model for ImageProcessingOptions.
type ImageProcessingOptions struct {
	Imageid        *string `json:"imageid,omitempty"`
	ProcessingType int     `json:"processingType"`
}

// JwtAccessToken Users JWT access token
type JwtAccessToken = string

// JwtRefreshToken Users JWT refresh token
type JwtRefreshToken = string

// LoginInput defines model for LoginInput.
type LoginInput struct {
	Email    openapi_types.Email `json:"email" validate:"email"`
	Password string              `json:"password" validate:"min=8,max=64"`
}

// ProcessingStatus defines model for ProcessingStatus.
type ProcessingStatus struct {
	Estimation *string `json:"estimation,omitempty"`
}

// RegisterInput defines model for RegisterInput.
type RegisterInput struct {
	Email    openapi_types.Email `json:"email" validate:"email"`
	Fullname string              `json:"fullname" validate:"min=2,max=64"`
	Password string              `json:"password" validate:"min=8,max=64"`
	Username string              `json:"username" validate:"min=2,max=64"`
}

// RegularAuthResponse defines model for RegularAuthResponse.
type RegularAuthResponse struct {
	// Email Authenticated users email
	Email *openapi_types.Email `json:"email,omitempty"`

	// Token Users JWT access token
	Token *JwtAccessToken `json:"token,omitempty"`
}

// BadRequest defines model for BadRequest.
type BadRequest = Error

// NotFound defines model for NotFound.
type NotFound = Error

// GoogleOAuthCbParams defines parameters for GoogleOAuthCb.
type GoogleOAuthCbParams struct {
	// Code Authorization code returned by Google OAuth2
	Code string `form:"code" json:"code"`

	// State Optional state parameter for CSRF protection
	State *string `form:"state,omitempty" json:"state,omitempty"`
}

// UploadImageMultipartBody defines parameters for UploadImage.
type UploadImageMultipartBody struct {
	Document *[]openapi_types.File `json:"document,omitempty"`
}

// LoginJSONRequestBody defines body for Login for application/json ContentType.
type LoginJSONRequestBody = LoginInput

// RegisterJSONRequestBody defines body for Register for application/json ContentType.
type RegisterJSONRequestBody = RegisterInput

// ProcessImageJSONRequestBody defines body for ProcessImage for application/json ContentType.
type ProcessImageJSONRequestBody = ImageProcessingOptions

// UploadImageMultipartRequestBody defines body for UploadImage for multipart/form-data ContentType.
type UploadImageMultipartRequestBody UploadImageMultipartBody